app-i18n/fcitx: block fcitx:4 and fcitx:5 each other
If we want to install both fcitx4 and fcitx5 at the same time,
we just need to adjust the location of one of the icon files.
But consider that fcitx4 and fcitx5 are registered in the same
im module. For example, their XMODIFIERS variables are both "@im=fcitx".
Running them at the same time is undefined behavior.

Bug: https://bugs.gentoo.org/760501
Signed-off-by: Yongxiang Liang <tanekliang@gmail.com>